+++ lib/src/storage/license.dart
@@ -3,7 +3,7 @@ import '../license_types.dart';
 /// Represents a raw license
 class License {
   /// Concrete license text
-  final String text;
+  final String? text;
 
   /// Determined type of license
   ///
@@ -14,12 +14,12 @@ class License {
 
   /// Create an empty license to still show
   /// the name of the package
-  License.notFound() : text = '?';
+  License.notFound() : text = null;
 
   @override
   String toString() {
     if (type != null) return '($type)\n$text';
 
-    return text;
+    return text.toString();
   }
 }